+++ tic-tac-toe/src/features/theme.tsx
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import type { RootState } from "../store";
+import type { RootState } from "../app/store";
 
 interface themeState {
     value: ''
