+++ tic-tac-toe/src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import {store} from './store'
+import {store} from './app/store'
 import {Provider} from 'react-redux'
